+++ CumulusMX/ImetStation.cs
@@ -525,6 +525,7 @@ namespace CumulusMX
 
 					var midnightraindone = luhour == 0;
 					var rollover9amdone = luhour == 9;
+					var snowhourdone = luhour == cumulus.SnowDepthHour;
 
 					for (int i = startindex; i < numrecs; i++)
 					{
@@ -544,17 +545,6 @@ namespace CumulusMX
 							cumulus.LogMessage("Processing logger data entry " + i + " for " + timestamp);
 
 							int interval = (int) (Convert.ToDouble(sl[INTERVALPOS], provider) / 60);
-							// Check for roll-over
-
-							if (hour != rollHour)
-							{
-								rolloverdone = false;
-							}
-
-							if (hour != 0)
-							{
-								midnightraindone = false;
-							}
 
 							if (sl[RELHUMAVGPOS].Length > 0)
 							{
@@ -694,7 +684,13 @@ namespace CumulusMX
 							// Add current data to the lists of web service updates to be done
 							cumulus.AddToWebServiceLists(timestamp);
 
-							if ((hour == rollHour) && !rolloverdone)
+							// Check for roll-over
+
+							if (hour != rollHour)
+							{
+								rolloverdone = false;
+							}
+							else if (!rolloverdone)
 							{
 								// do roll-over
 								cumulus.LogMessage("Day roll-over " + timestamp);
@@ -703,7 +699,12 @@ namespace CumulusMX
 								rolloverdone = true;
 							}
 
-							if ((hour == 0) && !midnightraindone)
+
+							if (hour != 0)
+							{
+								midnightraindone = false;
+							}
+							else if (!midnightraindone)
 							{
 								ResetMidnightRain(timestamp);
 								ResetSunshineHours(timestamp);
@@ -712,10 +713,36 @@ namespace CumulusMX
 							}
 
 							// 9am rollover items
-							if (hour == 9 && !rollover9amdone)
+							if (hour != 9)
+							{
+								rollover9amdone = false;
+							}
+							else if (!rollover9amdone)
 							{
 								Reset9amTemperatures(timestamp);
 								rollover9amdone = true;
+							}
+
+							// Not in snow hour, snow yet to be done
+							if (hour != cumulus.SnowDepthHour)
+							{
+								snowhourdone = false;
+							}
+							else if (!snowhourdone)
+							{
+								// snowhour items
+								if (cumulus.SnowAutomated > 0)
+								{
+									CreateNewSnowRecord(timestamp);
+								}
+
+								// reset the accumulated snow depth(s)
+								for (int j = 0; j < Snow24h.Length; j++)
+								{
+									Snow24h[j] = null;
+								}
+
+								snowhourdone = true;
 							}
 						}
 						catch (Exception E)